Refactoring model, generated structures for Go

+++ cars/model/image.go
@@ -3,7 +3,7 @@ package model
 type Image struct {
 	Id int `gorm:"primary_key"`
 	Encoded64Image string
-	CarId    uint
+	CarId    int64
 }
